feat (createSubmission) : change date reference for recurringDate loop.
Previously, startDate was used as the reference for eventDate. However, this would lead to UI confusion as a an existing event's period would be rendered as existing Start Date till End Date. To be clearer, eventDate field was created. This field will be reference for calendar rendering logic, while startDate will be used to reference actual start to end period